+++ firmware/CHANGELOG.md
@@ -34,6 +34,7 @@ All notable user-facing or behavior-altering changes will be documented in this 
  - Toyota 4-1 VVT trigger
  - use extra RAM for lua if your microRusEFI unit has stm32f42x microcontroller 
  - VVT activation delay #3443
+ - ECU uptime into critical error message 
 
 ## March 2022 Release - "Day 22"